Add files via upload
Firstly, we uploaded the data of Seoul Bike Data. Then, make sure every element is in the right position and form, for example, ‘Date’ in the correct date form. Extract the month and year of the date respectively and the amount of rented bike to form the line chart. Combine the month and year to the same category, in order to draw the chart according to the chronological order. Create the chart and y-axis(ax1). Draw the line charts (the lower one) to show the daily amount, and then draw the bar charts on the y-axis. Create the second y-axis(ax2) and use the x-axis with the first ax1. Draw the line charts (the upper one) to show the monthly amount, and then put the upper line chart on ax2. Set x-axis as the date form, and the topic and the label of x-axis. Rotate the label of x-axis 45 degrees. Show the icons.